all: show 'DRAFT' watermark on the first page
Unconditionally show a 'DRAFT' watermark only on the first page of each
document.

I would have liked to make this conditional based on the BUILD_RELEASE=*
variable, but could not figure out how to switch off a .sty code segment based
on command line parameters.

Since we only seem to render all manuals in DRAFT mode all the time, it seems
reasonable to include the DRAFT watermark always, requiring manual .sty editing
if anyone wanted to build a non-DRAFT version.

In the makefile, add an echo to say so in case of non-draft builds.

Change-Id: Ia71795481c2467aab11ab344517df74775b2a852