prepare for autotools: add git-version-gen
Add the same version of git-version-gen, which most other Osmocom
repositories (libasn1c, libosmocore, libosmo-netif, libosmo-sccp, ...)
are using, so they behave the same.

git-version-gen generates a version string derived from the git
repository, or from a .tarball-version file when running outside of a
git directory. Unfortunatelly it seems to be impossible to use
git-version-gen from autoconf without adding it to the source tree.

Autotools will be used to install the shared manual files into the
system in a way that is consistent with the other Osmocom repositories:

$ autoreconf -fi
$ ./configure --prefix=/usr/local
$ make
$ make install

Project repositories will check with pkg-config if osmo-gsm-manuals is
installed (like any other shared Osmocom component), when a new
configure flag for building the manuals is passed. This will also work
with distribution's package managers, for instance we could have a
osmo-gsm-manuals-dev package in Debian on which the other packages
would have a build-time dependency to build the UNIX man pages
(OS#3386).

(moving manuals to project repositories 4/19)

Related: OS#3385
Change-Id: I34b0d20046417179250a5065574e2bc303889cf6