vty_reference_combine.sh: chmod xsltproc output
Force the xsltproc output file to be writable, if it exists already.
This is needed for 'make distcheck': xsltproc gives output files the
same permissions as input files. We need to change the output file in
multiple iterations in vty_reference_combine.sh, hence it must be
writable.

(moving manuals to project repositories 15/19)

Related: OS#3385
Change-Id: I693e12ee15665d01cbc50187e6cb3ca74ad0356a