debian: add patch for debian 8
Add a patch file that removes the asciidoc-dblatex dependency in
debian/control, so the package builds for debian 8. In debian 8, the
files are all in the asciidoc package, which does get installed.

The jenkins scripts will copy the source package, apply the patch, and
upload it separately to OBS [1]. A similar debian 8 patch exists for
osmo-trx.

I have also tried to do it without a patch, by changing the OBS
configuration. It is possible to bypass the dependency check from OBS
and get the build started, but there's an additional check by
dpkg-checkbuilddepends and this one does not pass. So we really need
the patch method.

[1]: Change-Id I3570599ede51b974d350064f44f77e360fafd8b0 (osmo-ci)

Related: OS#3899
Change-Id: I5b9575ceb1141961e570643a5755a2bd6b6a4254