GSUP: document inter-MSC handover additions
Message formats of the new messages look mostly the same (IMSI,
Message Class, Source Name, Destination Name, AN-APDU). That is, because
AN-APDU is storing results, error reasons etc. This can be seen clearly
in osmo-msc.git:
* src/libmsc/msc_a_remote.c:msc_a_remote_fsm_communicating()
* src/libmsc/msc_i_remote.c:msc_i_remote_fsm_ready()

The message squence charts in the E Procedures section are directly
based on Neels' interMSC_HO_GSUP_msgs.txt [1].

It seems that using AN-APDU made some other new IEs redundant: RR Cause,
BSSAP Cause, Session Management Cause had been added to GSUP for the MSC
handover, and are documented now, but they are currently not used in
osmo-msc.git. The new message OSMO_GSUP_MSGT_E_ABORT is not used either,
so I left a stub for it in the message format section.

I mentioned in the Source Name IE section, that source and destination
names are sent as nul-terminated strings. This is for legacy reasons,
Neels wrote a nice summary in the commit message of [2].

[1] https://osmocom.org/attachments/3720/interMSC_HO_GSUP_msgs.txt
[2] Change-Id: I9ca8c9eef104519ed1ea46e2fef46dcdc0d554eb (osmo-msc)

Related: OS#3774, OS#3619
Change-Id: I6b9f23d08cfe53c8b77f51c6afb900c2badc9e2c