debian: fix empty debian package
Replace '--with autotools_dev' with '--with autoreconf' in
debian/rules, so a ./configure file gets generated from configure.ac,
and the debian package scripts don't generate an empty package.

I did not notice this earlier, because when building the package locally
with dpkg-buildpackage, it was not empty. (That is probably, because my
local dir already had the configure script generated.)

Related: OS#3899
Change-Id: I483136cd28395ae8fc2c112cf659ee83255afe87