Makefile.asciidoc.inc: always run verbose asciidoc
a2x is a "toolchain manager" for asciidoc, that runs the asciidoc
command internally. Replace the previous logic of first running a2x
without any verbose flags, and on failure running asciidoc standalone,
with a verbose flag, to print the errors the occured. Always run a2x
with -vv, so it runs asciidoc in verbose mode.

This prevents losing the error message in case asciidoc or a2x fail
when running with a2x, but then asciidoc succeeds when running
standalone.

Related: OS#4132
Change-Id: I2cbd2985645ecdecc7851849be6a9a95d4b3246e