TRXDv2: add primary/shadow classification for VAMOS PDUs
Introduce a new flag (1-bit field) indicating whether a VAMOS PDU
is for a _primary_ ('0'B) or for a _shadow_ ('1'B) timeslot.  This
field shall be present only in the Uplink VAMOS PDUs, which unlike
the Downlink (AQPSK) PDUs are sent separately.  This will simplify
routing of such PDUs in osmo-bts.

Change-Id: Ie00a12824938c6a65480bde1af364dcc9febb8f7
Related: SYS#4895, OS#4941, OS#4006